requirements.txt
backports.functools-lru-cache==1.4
configobj==5.0.6
cycler==0.10.0
decorator==4.0.11
et-xmlfile==1.0.1
future==0.16.0
imageio==2.1.2
jdcal==1.3
json-tricks==3.11.0
lxml==4.1.1
matplotlib==2.1.1
moviepy==0.2.3.2
numpy==1.13.3
openpyxl==2.4.9
pandas==0.22.0
Pillow==5.0.0
PsychoPy==1.85.3
pyglet==1.3.0
PyOpenGL==3.1.0
pyparsing==2.2.0
pypiwin32==219
python-dateutil==2.6.1
pytz==2017.3
scipy==1.0.0
six==1.11.0
tqdm==4.11.2
wxPython==4.0.0b2
